Image-to-Text-Server/pyproject.toml
[project]
name = "image-to-text-server"
version = "0.1.0"
description = "Image to Text Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.0",
    "torch>=2.6.0",
    "transformers>=4.52.4",
    "accelerate>=1.7.0",
    "bitsandbytes>=0.46.0",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
]
